Move layer extensions to main dir
