--- docs/api/experimental/traits.md
@@ -1,4 +1,0 @@
-# lonboard.experimental.traits
-
-::: lonboard.experimental.traits.PointAccessor
-::: lonboard.experimental.traits.GetFilterValueAccessor
