+++ docs/api/traits.md
@@ -5,3 +5,7 @@
 ::: lonboard.traits.ColorAccessor
 
 ::: lonboard.traits.FloatAccessor
+
+::: lonboard.traits.PointAccessor
+
+::: lonboard.traits.GetFilterValueAccessor
