+++ lonboard/experimental/__init__.py
@@ -5,8 +5,3 @@ unexpected behavior when using them.
 """
 
 from ._layer import ArcLayer, TextLayer
-from .layer_extension import (
-    BrushingExtension,
-    CollisionFilterExtension,
-    DataFilterExtension,
-)
